requirements.txt
pygad
scipy
setuptools
benchmark_functions
opfunu~=1.0.1
optuna~=4.1.0
sqlalchemy
optuna-dashboard
pytest~=8.3.4
numpy~=2.2.0
matplotlib~=3.10.0
datetime~=5.5
hydra-core~=1.3.2
omegaconf~=2.3.0
tqdm~=4.67.1